rename whitesource command: move to own file
